pyproject.toml
[tool.poetry]
name = "Nikita-Illarionov_gendiff"
version = "0.2.5"
description = ""
authors = ["Nikita <Nikita.Clark@yandex.ru>"]
packages = [
 { include = "gendiff" },
]

[tool.poetry.dependencies]
python = "^3.6"
PyYaml = "^5.3.1"

[tool.poetry.dev-dependencies]
flake8 = "^3.8.3"
pytest = "^6.1.0"
coverage = "^5.3"
pytest-cov = "^2.10.1"

[tool.poetry.scripts]
gendiff = 'gendiff.scripts.gendiff:main'

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
